feat: "CenteredImg" 제작
- 아무 퍼블릭 이미지나 가운데 크게 띄워줄 수 있음
