--- client/src/components/Empty/index.ts
@@ -1,1 +1,0 @@
-export { Empty } from './Empty'
